pyproject.toml
[project]
name = "dcs-world-schema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.2",
    "ruamel-yaml>=0.18.10",
    "ruff>=0.11.9",
    "yamllint>=1.37.1",
]
